♻️ extracte api_base from api 🏳 direct use asgi
- using asgi directly made the app to be built
on each petition and we are not php
- extracted a new setup (indeed two, because
statics are meant to be used at the end
to avoid shadow every path.
