+++ backend/api.py
@@ -1,28 +1,11 @@
+import os
 from fastapi import FastAPI
-from fastapi.staticfiles import StaticFiles
-from pathlib import Path
 from dotenv import load_dotenv
-from . import __version__ as version
+from .api_base import setup_base, setup_statics
 from .authlocal import setup_authlocal
 from .api_business import setup_business
 
-def setup_base(app):
-    @app.get('/api/version')
-    def apiVersion():
-        return dict(
-            version = version,
-        )
-
-
-def setup_statics(app):
-    # IMPORTANT!
-    # This should be called the last of the setups
-    # since it is registered at / and will catch all routes
-    packagedir = Path(__file__).parent
-    distpath = packagedir/'dist'
-    app.mount("/", StaticFiles(directory=distpath, html=True), name="ui")
-
-async def app(scope, receive, send):
+def setup():
     load_dotenv()
     app = FastAPI()
     setup_base(app)
@@ -31,6 +14,7 @@ async def app(scope, receive, send):
     setup_business(app)
     setup_statics(app)
     [print(r) for r in app.routes]
-    return await app(scope, receive, send)
+    return app
 
+app = setup()
